+++ src/room/Room.ts
@@ -1407,6 +1407,10 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
   private handleSpeakersChanged = (speakerUpdates: SpeakerInfo[]) => {
     const lastSpeakers = new Map<string, Participant>();
     this.activeSpeakers.forEach((p) => {
+      const remoteParticipant = this.remoteParticipants.get(p.identity);
+      if (remoteParticipant && remoteParticipant.sid !== p.sid) {
+        return;
+      }
       lastSpeakers.set(p.sid, p);
     });
     speakerUpdates.forEach((speaker) => {
@@ -1514,14 +1518,14 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
   bufferedSegments: Map<string, TranscriptionSegmentModel> = new Map();
 
   private handleTranscription = (
-    remoteParticipant: RemoteParticipant | undefined,
+    _remoteParticipant: RemoteParticipant | undefined,
     transcription: TranscriptionModel,
   ) => {
     // find the participant
     const participant =
       transcription.transcribedParticipantIdentity === this.localParticipant.identity
         ? this.localParticipant
-        : remoteParticipant;
+        : this.getParticipantByIdentity(transcription.transcribedParticipantIdentity);
     const publication = participant?.trackPublications.get(transcription.trackId);
 
     const segments = extractTranscriptionSegments(transcription);